Changed sign of input Neumann stress for MPSA.
If all boundary normals point outward there should be no change in behaviour.
For boundary normals that point inwards the neumann stress should be flipped.

The new behaviour does not require the user to know anything about the normals
when he or she sets the stress boundary condition. Setting the stress on a face
to [1,1] will result in a force up and to the right for both outward pointing
normals and inward pointing normals. Previously, this would result in a force
pointing down to the left for inward pointing normals (due to T_ik = - T_ki).