Fixed bug in partition of grid where node ordering was shuffled.
It is very important to keep the node ordering of the nodes of each face.
If face i has nodes [5,3,4], that is;
face_nodes.indices[face_nodes.indptr[i]:face_nodes.indptr[i+1]] = [5,3,4],
this is note the same as it having the nodes [3, 4, 5]. This is because
the nodes are assumed to be ordered counterclockwise (or clockwise, I don't remember),
for many algorithms. If you slide face_nodes[:, i] you will get out the sorted order;
[3, 4, 5], instead of what you want: [5,3,4]. Have a look at pp.utils.sparse_mat for
utility functions to manipulate the sparse matrices without changing the indices ordering.