Changed displacment reconstruction from subhfno to subfno.
Earlier each gradient gave one displacment on each of its subfaces.
Since we have a matching displacment, the displacement from the two
gradients sharing a subface will be equal. Therefore we changed
the displacement reconstruction to be given on unique subfaces, that
is, subfno. We now take an average of the two sides, could also have
just picked one of them