Changed continuity_point back to mpf(s)a_eta. Also moved discretization params to Paramter dictionary.
The naming change of eta does not belong here. If the eta name is changed, this
has to be done on a module level (not just at the outer most level).

The natural choice now is that all parameters for the discretization now belongs in the
Paramert dictionary, not the outer dictionary