feat: Method to run a convergence study
the method run_model_for_convergence_study allows any subclass of ContactMechanics to be used as a model.
The run_model_method is typically pp.run_time_dependent_model.
For a given pp.FractureNetworkNd (N=2,3), we run a convergence study over a defined set of variables.
This is essentially a convenient wrapper for computing errors over the gb_coarse_fine_cell_mapping method