fix: bug in  find error in log (gmsh_interface.py)
This line incorrectly raises a ValueError if any part of the log line,
not only the log level, contains the name "Error".
For example, an info message will contain the "in_file" path. If that path
contains the word "Error", this method will raise.

Note: This fix is not by any means viable in the long run, given different logginf formats,
we should discuss how to do this properly.