ENH: Extend import such that it also works for complex grids.
Complex grids here mean grids with cells of different cell type,
typically appearing for polyhedral grids. Currently one test is
still failing. This is due to a bug in export_polyhedra_3d, which
will be fixed in a separate PR - this issue should be fixed
when this PR is accepted and the corresponding branch is merged.