DOC: Updated md_grid and mortar_grid
Some comments:

I am not entirely sure whether I got the shapes of all arrays correct, so these should be double-checked.

Line 398 of the updated mortar_grid: Not sure if this ValueError is correctly explained.

I think faces and cells should be swapped in some of the explanations (e.g line 339 in the updated md_grid), but again, not 100% sure.

The type "list[pp.Grid]" will be displayed as "list[porepy.grids.grid.Grid]" on the html page (same with tuple[pp.Grid]). I'm guessing we do not want this, as it becomes much harder to read? I do not know how to fix this, however.