DEP/MAINT: Deprecate `StoreResults` and use `DataSavingMixin` instead.
The `StoreResults` class is now deprecated. Instead, a modification of `DataSavingMixin` (i.e., `ModifiedDataSavingMixin`) is employed. To store the results of a verification, the method `store_data()` provided by `ModifiedDataSavingMixin` is called in the `after_nonlinear_loop()` method. If the current time matches a time from `stored_times`, a `SaveData` object (provided by a new dataclass with the same name) is appended to the `results` list.