BUG: Cartesian fracture meshing with intersections
In Cartesian meshing of fractured domains, the fractures are snapped to
the underlying mesh. The subsequent identification of intersections did
not consider the snapped fractures, leading to intersection points
being identified away from the grid lines, which caused errors.

Also modified one of the meshing tests slightly so that it would have
picked up the bug.