TST: Update reference values in poromechanics test
The old poromechanics model had the wrong scaling of the consistency
('stability') term in the poromechanics model: The scaling was off by a
factor of alpha (the Biot coefficient) see Nordbotton 2016 SIAM Num An,
Remark 13 for details.

Before updating the reference values it was verified that, after adding
the alpha factor in front of the consistency term, the discretization
matrices produced by the new and old models are the same.