MAINT: Rearranged init methods for several ad operators
With the new graph-based representation of operator trees generates a networkx graph
for all nodes as part of the superclass' init.
This leads to several calls of _key() for the individual subclasses.
The call will fail unless data fields needed in _key() have already been populated.

There should be no downsides in the rearrangement, fingers crossed.