TST: Sort lists of tested and testable methods in fluid mass balance
For some reason, one of the lists of tested and testable methods has been shuffled. This is potentially because the order in which the constitutive laws are accessed has changed
(due to changes done during introduction of the MatrixSlicer, though EK does not understand exactly how).

There is no reason why the two lists should be arranged identically, so sorting should be okay.