TST: Helper methods for consistent cell-face relations in tests
In some tests, the cell-face relation of a grid is not important, and
grids have been constructed with only positive values in this matrix.
This is no longer permissible, and the tests therefore raised an error.

This commit introduces a helper method to fix such inconsistencies by an
arbitrary assignment of values -1 at suitable places in the cell-face
relation matrix.