TST: Update reference array for mpfa pressure reconstruction
When the face-node mapping in grid construction was changed
(commit d37426105), the ordering of the sub-faces in mpfa also changed
(think: The upper and lower part of a face had their ordering swapped).

It was verified that the change to the reference array involved
swapping rows for those faces where the ordering changed.